Allow callers to specify a block to be run on completion of dismissal animation.
This allows a workaround for the issue where after dismissal, the status bar returns to having
black content, regardless of whether white content should be shown instead. The caller can now
call setNeedsStatusBarAppearanceUpdate in the completion handler to ensure that the system
always draws the correct colours in the status bar.