Fix a bug where status bar notification could only be used once.
This change re-uses self.statusBarView as well as self.notificationLabel, but creates a new UIWindow
for each notification. The crucial part, though, is to reset the label’s transform to CGAffineTransformIdentity.
If you were using CWNotificationAnimationStyleBottom, the dismissal animation scaled the height of the label
down to zero without resetting it, which meant that while it was technically being reused properly, it wouldn’t
display any pixels since it had zero height.